+++ Makefile.toml
@@ -65,6 +65,10 @@ args = [
     "**/src/**",
     "--ignore",
     "**/.cargo/**",
+    "--ignore",
+    "**/target/coverage/**",
+    "--ignore",
+    "**/target/debug/**",
     "-o",
     "./target/coverage/lcov.info",
 ]
@@ -86,6 +90,10 @@ args = [
     "**/src/**",
     "--ignore",
     "**/.cargo/**",
+    "--ignore",
+    "**/target/coverage/**",
+    "--ignore",
+    "**/target/debug/**",
     "-o",
     "./target/coverage/coverage.md",
 ]
