+++ crates/service/src/board/persist.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use async_graphql::connection::{Connection, Edge};
 use indoc::indoc;
 use tracing::instrument;
@@ -7,7 +10,7 @@ use crate::{
     account::CurrentAccount,
     persist::Persist,
     prelude::*,
-    query::{values_table, OpaqueCursor, PaginationInput, PaginationOptions, ResultSlice},
+    query::{OpaqueCursor, PaginationInput, PaginationOptions, ResultSlice},
 };
 
 pub struct BoardPersist<'a> {
@@ -82,9 +85,20 @@ impl<'a> BoardPersist<'a> {
     }
 
     #[instrument(skip_all)]
-    pub async fn update(&self, _id: &str, _update: UpdateBoard) -> Result<Board> {
-        // self.persist.db().update((TABLE_NAME, id)).patch(patch_op)
-        todo!()
+    pub async fn update(&self, id: &str, update: UpdateBoard) -> Result<Board> {
+        // TODO: check config to see if anon users can update boards
+        // TODO: check perms to see if authd user can update boards
+
+        let board = if let Some(update) = update.into_update((TABLE_NAME, id).into()) {
+            self.persist.db().query(update).await?.take(0)?
+        } else {
+            self.get(id).await?
+        };
+
+        match board {
+            Some(board) => Ok(board),
+            None => Err(Error::NotFound(format!("board {id}"))),
+        }
     }
 
     #[instrument(skip_all)]
@@ -125,7 +139,7 @@ impl<'a> BoardListRequest<'a> {
 
         let query = srql::SelectStatement {
             expr: srql::Fields::all(),
-            what: values_table(TABLE_NAME),
+            what: srql::values_table(TABLE_NAME),
             order: srql::Orders(order.into_iter().collect()).into(),
             cond,
             limit,
@@ -146,204 +160,6 @@ impl<'a> BoardListRequest<'a> {
             .collect();
 
         Ok(connection)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use pretty_assertions::assert_eq;
-    use std::{collections::VecDeque, pin::pin};
-
-    use super::{testing::BoardTestData as _, *};
-    use crate::{account::testing::*, query::testing::Paginator};
-
-    #[tokio::test]
-    async fn test_create() {
-        let data = TestData::new().await;
-        let board_persist = data.board();
-
-        let board = CreateBoard {
-            handle: Some("test".into()),
-            name: Some("Test".into()),
-            description: Some("Test".into()),
-        };
-
-        let res = board_persist.create(board).await;
-        println!("{res:?}");
-        assert!(res.is_ok());
-
-        let res = res.unwrap();
-        assert_eq!(res.handle, "test");
-        assert_eq!(res.name, Some("Test".to_owned()));
-        assert_eq!(res.description, Some("Test".to_owned()));
-    }
-
-    #[tokio::test]
-    async fn test_get() {
-        let data = TestData::new().await;
-        let board_persist = data.board();
-        let board = data.generate_board().await;
-
-        let res = board_persist.get(&board.id.id.to_raw()).await;
-        println!("{res:?}");
-        assert!(res.is_ok());
-
-        let res = res.unwrap();
-        assert!(res.is_some());
-
-        let res = res.unwrap();
-        println!("{res:?}");
-        assert_eq!(res.id, board.id);
-        assert_eq!(res.creator_id, board.creator_id);
-        assert_eq!(res.handle, board.handle);
-        assert_eq!(res.name, board.name);
-        assert_eq!(res.description, board.description);
-    }
-
-    #[tokio::test]
-    async fn test_get_handle() {
-        let data = TestData::new().await;
-        let board_persist = data.board();
-        let board = data.generate_board().await;
-
-        let res = board_persist.get_by_handle(&board.handle).await;
-        println!("{res:?}");
-        assert!(res.is_ok());
-
-        let res = res.unwrap();
-        assert!(res.is_some());
-
-        let res = res.unwrap();
-        println!("{res:?}");
-        assert_eq!(res.id, board.id);
-        assert_eq!(res.creator_id, board.creator_id);
-        assert_eq!(res.handle, board.handle);
-        assert_eq!(res.name, board.name);
-        assert_eq!(res.description, board.description);
-    }
-
-    #[tokio::test]
-    async fn test_duplicate_handle() {
-        let data = TestData::new().await;
-        let board_persist = data.board();
-        let board = data.generate_board().await;
-
-        let create = CreateBoard {
-            handle: Some(board.handle),
-            name: Some("Test".into()),
-            description: Some("Test".into()),
-        };
-
-        let res = board_persist.create(create).await;
-        println!("{res:?}");
-        assert!(res.is_err());
-
-        let res = res.unwrap_err();
-        println!("{res:?}");
-        assert_eq!(res, Error::UnavailableIdent);
-    }
-
-    #[tokio::test]
-    async fn test_anon_create_default_fail() {
-        let data = TestData::new().await;
-        let board_persist = data.board();
-
-        let board = CreateBoard {
-            handle: None,
-            name: Some("Test".into()),
-            description: Some("Test".into()),
-        };
-
-        let res = board_persist.create(board).await;
-        println!("{res:?}");
-        assert!(res.is_err());
-
-        let res = res.unwrap_err();
-        println!("{res:?}");
-        assert_eq!(res, Error::MissingIdent);
-    }
-
-    #[tokio::test]
-    async fn test_authed_create_default() {
-        let (data, acc) = TestData::with_user().await;
-        let board_persist = data.board();
-
-        let board = CreateBoard {
-            handle: None,
-            name: Some("Test".into()),
-            description: Some("Test".into()),
-        };
-
-        let res = board_persist.create(board).await;
-        println!("{res:?}");
-        assert!(res.is_ok());
-
-        let res = res.unwrap();
-        println!("{res:?}");
-        assert_eq!(res.handle, acc.user_id);
-        assert_eq!(res.creator_id, Some(acc.id));
-    }
-
-    #[tokio::test]
-    async fn test_forward_pagination() {
-        let (data, _) = TestData::with_user().await;
-        let test_data = data.generate_boards(50).await;
-        let board_persist = pin!(data.board());
-
-        let mut results = vec![];
-        let mut paginator = Paginator::new(|cursor| async {
-            board_persist
-                .list()
-                .with_pagination(PaginationInput::new().forward(10).set_after(cursor))
-                .execute()
-                .await
-        });
-
-        while let Some(res) = paginator.next().await {
-            assert!(res.is_ok());
-
-            let res = res.unwrap();
-            assert_eq!(res.len(), 10);
-
-            results.extend(res);
-        }
-
-        assert_eq!(results.len(), test_data.len());
-        assert_eq!(results, test_data);
-    }
-
-    #[tokio::test]
-    async fn test_backward_pagination() {
-        let (data, _) = TestData::with_user().await;
-        let test_data = data.generate_boards(50).await;
-        let board_persist = pin!(data.board());
-
-        let mut results = VecDeque::new();
-        let mut paginator = Paginator::new(|cursor| async {
-            board_persist
-                .list()
-                .with_pagination(PaginationInput::new().backward(10).set_before(cursor))
-                .execute()
-                .await
-        })
-        .reversed();
-
-        // Because reverse pagination still keeps the items in each page in
-        // the same order as forward pagination, but the pages are in reverse
-        // order, we need to collect the pages in reverse order to get the
-        // correct order of items while keeping the item order.
-        while let Some(res) = paginator.next().await {
-            assert!(res.is_ok());
-
-            let res = res.unwrap();
-            assert_eq!(res.len(), 10);
-
-            results.push_front(res);
-        }
-
-        let results: Vec<_> = results.into_iter().flatten().collect();
-        assert_eq!(results.len(), test_data.len());
-        assert_eq!(results, test_data);
     }
 }
