--- crates/service/src/query.rs
@@ -1,3 +1,0 @@
-mod pagination;
-
-pub use pagination::*;
